requirements.txt.txt
streamlit
opencv-python
numpy
easyocr
imutils
Pillow